requirements.txt
pandas
tqdm
matplotlib
dgl
numpy
torchsummary